applications/version-demo-server-py/requirements.txt
aiokafka
asyncio
boto3
fastavro
python-schema-registry-client==1.8.2
websockets
